+++ agent/lm_agent/workload_managers/slurm/cmd_utils.py
@@ -120,7 +120,7 @@ async def make_booking_request(lbr: LicenseBookingRequest) -> bool:
     logger.debug(f"lbr: {lbr}")
 
     resp = await async_client().put(
-        "/api/v1/booking/book",
+        "/lm/api/v1/booking/book",
         json={
             "job_id": lbr.job_id,
             "features": features,